requirements.txt
Pillow>=10.0.1
pygame==2.2.0